refactor(test): add acompat module
