+++ tests/pool/test_pool_null.py
@@ -12,7 +12,7 @@ from psycopg.pq import TransactionStatus
 from psycopg.rows import class_row, Row, TupleRow
 from psycopg._compat import assert_type
 
-from ..utils import Event, sleep, spawn, gather, is_async
+from ..acompat import Event, sleep, spawn, gather, is_async
 from .test_pool_common import delay_connection, ensure_waiting
 
 try:
