+++ tests/test_cursor_common_async.py
@@ -13,7 +13,8 @@ from psycopg import sql, rows
 from psycopg.adapt import PyFormat
 from psycopg.types import TypeInfo
 
-from .utils import gc_collect, raiseif, aclosing, alist, anext
+from .utils import gc_collect, raiseif
+from .acompat import aclosing, alist, anext
 from .fix_crdb import crdb_encoding
 from ._test_cursor import my_row_factory, ph
 from ._test_cursor import execmany, _execmany  # noqa: F401
