+++ tests/test_pipeline_async.py
@@ -9,7 +9,7 @@ import psycopg
 from psycopg import pq
 from psycopg import errors as e
 
-from .utils import is_async, anext
+from .acompat import is_async, anext
 
 pytestmark = [
     pytest.mark.pipeline,
